move stuff to windows.rs. closes #105
